REF: misplaced arithmetic tests (#32912)
